ai_doctor_chatbot/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
requests
python-dotenv